build: npm 배포 설정
