+++ hooks/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "hooks",
+  "name": "@jae-o/hooks-module",
   "version": "0.0.1",
-  "description": "npm modules",
+  "description": "hooks module",
   "type": "module",
   "types": "dist/index.d.ts",
   "main": "dist/index.js",
@@ -41,5 +41,8 @@
     "typescript": "^5.8.3",
     "vite": "^6.2.6",
     "vite-plugin-dts": "^4.5.3"
+  },
+  "publishConfig": {
+    "access": "public"
   }
 }
