--- hooks/src/cardNumber/constants.ts
@@ -1,8 +1,0 @@
-export const CARD_NUMBERS_KEY = {
-  part1: 'part1',
-  part2: 'part2',
-  part3: 'part3',
-  part4: 'part4',
-};
-
-export type CardNumbersKey = keyof typeof CARD_NUMBERS_KEY;
