--- hooks/src/expiryDate/constants.ts
@@ -1,6 +1,0 @@
-export const EXPIRY_DATE_KEY = {
-  month: 'month',
-  year: 'year',
-};
-
-export type ExpiryDateKey = keyof typeof EXPIRY_DATE_KEY;
