+++ hooks/src/lib/constants.ts
@@ -1,3 +1,17 @@
+export const CARD_NUMBERS_KEY = {
+  part1: 'part1',
+  part2: 'part2',
+  part3: 'part3',
+  part4: 'part4',
+};
+export type CardNumbersKey = keyof typeof CARD_NUMBERS_KEY;
+
+export const EXPIRY_DATE_KEY = {
+  month: 'month',
+  year: 'year',
+};
+export type ExpiryDateKey = keyof typeof EXPIRY_DATE_KEY;
+
 export const CARD_NUMBER_ERROR_TYPES = {
   notNumber: 'notNumber',
   invalidLength: 'invalidLength',
