+++ hooks/tsconfig.app.json
@@ -20,7 +20,9 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
-    "noUncheckedSideEffectImports": true
+    "noUncheckedSideEffectImports": true,
+    "declaration": true,
+    "declarationDir": "dist"
   },
-  "include": ["src"]
+  "include": ["src", "src/lib"]
 }
