+++ hooks/vite.config.ts
@@ -1,26 +1,32 @@
-import * as path from "path";
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react-swc";
-import dts from "vite-plugin-dts";
+import * as path from 'path';
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react-swc';
+import dts from 'vite-plugin-dts';
 
 export default defineConfig({
   build: {
     lib: {
-      entry: path.resolve(__dirname, "src/lib/index.ts"),
-      name: "index",
-      fileName: "index",
+      entry: path.resolve(__dirname, 'src/lib/index.ts'),
+      name: 'index',
+      fileName: 'index',
     },
     rollupOptions: {
-      external: ["react"],
+      external: ['react'],
       output: {
         globals: {
-          react: "React",
+          react: 'React',
         },
       },
     },
     commonjsOptions: {
-      esmExternals: ["react"],
+      esmExternals: ['react'],
     },
   },
-  plugins: [react(), dts()],
+  plugins: [
+    react(),
+    dts({
+      include: ['src/lib'],
+      tsconfigPath: './tsconfig.app.json',
+    }),
+  ],
 });
